Revert "Fix e2e tests"
This reverts commit 4393fb4951fa9ae5217c3da067cb5a1cc8ad0769.
