+++ src/applications/vaos/tests/e2e/workflows/direct-schedule-workflow/calendar-dead-ends.cypress.spec.js
@@ -14,12 +14,12 @@ import PreferredDatePageObject from '../../page-objects/PreferredDatePageObject'
 import TypeOfCarePageObject from '../../page-objects/TypeOfCarePageObject';
 import VAFacilityPageObject from '../../page-objects/VAFacilityPageObject';
 import {
+  mockAppointmentsGetApi,
   mockClinicsApi,
   mockEligibilityApi,
   mockEligibilityCCApi,
   mockEligibilityDirectApi,
   mockEligibilityRequestApi,
-  mockExistingAppointments,
   mockFacilitiesApi,
   mockFeatureToggles,
   mockSchedulingConfigurationApi,
@@ -36,7 +36,7 @@ describe('VAOS direct schedule flow - calendar dead ends', () => {
   beforeEach(() => {
     vaosSetup();
 
-    mockExistingAppointments(true);
+    mockAppointmentsGetApi({ response: [] });
     mockFeatureToggles();
     mockVamcEhrApi();
   });
